ready to build the parser
parser is just a pile of state machine, different constructs may
need different ones. the goal is to keep code short as possible,
correct and a little bit consideration for tree walking. try to
reuse state machines of the same constructs.

since the recursion happens inside the transfer-condition function,
in other words, when the transfer condition is a **non-terminal**
it needs further parsing to tell.

use global static ast_node ptr to get the result, use global static
token_buffer ptr to ask for tokens when parsing is non-trivial.

design the ast_node family, make them harmony.

impl the algo the parse expressions and the infamous C's modifiers
to the base type.

impl the memory pool for token and token*, -- not urgent.