NOTICE, from here goes the state-table refactor
reasons:
currently the state-table 's flexibility even sucks the
tokenizer, how can I expect it works for parser?
suck1 : use of bit feild kill or waste time of debuging

suck2 : action and flag have no more fucking use than
let you swith them. but only 3 kinds of actions will
never be enough.

suck3 : error handling depends on the previous state, the
table itself run into null_state if it failed. doing error
guessing on failed-previous-state is no good. why not just
determine them when defining the state-table?

suck4 : the add_* api is short but useless, we can use more
context-dependent way to save typing, like a configure on
the action and `from` state.

suck5 : the table assumes that reversed is a essential property
of how to transfer, but in fact it is not. there should be feilds
for the user to assign **his** data for transfer-judgement usage.

suck6 : the state-table can not manage its state, which is hard-
coding now, which is a real suck. state should be alloc unlimitedly,
table should grow itself.

another straw is , after I hard coded the transfers, I fuckingly
still have to hard-code the states and conditions! why should not
they alloc themselves!

after the machine is built, who can care what is in in?
may be for debuging purposes I can give name to them, but
in real code, I WILL NEVER WANT TO SEE their fucking long names!
let it handle itself.

still another reason is, the vast variety the catagory and subtypes
of tokens, non-terminals, types. currently I fuck of a beautiful way
to tackle this: there's level structure in tokenizer-parser-typesystem
codegen: each one can see 2 part of info about a construct: one part is
closely related to the current level, another part is of no use to it,
but for further analysis.
for the tokenizer, it sees one kind of token having a varlen buffer, yet
anthor kind -- punctuation, operator, keyword, have only symbol names.
but what it can also see, is, punctuation has many subtype, comma, dot,
operator has many subtype: add, sub, and, or. keyword has if , for ,...

there must be more level in the parser's eyes. if I failed to think of
the correct way to handle these, big refactor will continue