new api for state-table
made it a little bit like xml or so style.
config_**(from, to, action, condion) begin
an indent, or a scope, util the next call to
the same config_** with diff arg, the following
add_**(to,from) with take these configuration,
use a global data to track that.

but wasn't it too big a change?