fix the exprstmt bug
exprstmt := [expr] ;
the problem is with null. if expr is absent, only if we
successfully parse a ; can we push null to the node stack,
thus the optional_expr is not suitable here since it push
a null not depending on the tailing ;
but for the 'for' we can still use optional_expr.