fixup assign := condition | unary assign_op condition
conflict, through rewriting the rule into this form:
condition := (unary assign_op) * condition
zero or more "unary assign_op" sequence followed by "condition".
the algo is, parse and reduce the left most "unary assign_op"
as more as possible, when the current token is no longer "assign_op",
leave the "unary" on the stack and call "condition" to reduce it to
condition. to prevent unary(called by condition) ignores the unary on the top of stack
and consumes tokens, a "unary_on_top" bool is set inside context to
tell it look at the top