code the tostring part of Utillib.JSON
and updated Utillib.String.
added more a replace_last method, since
we need to "regret" the last char of a
string.
consider `[1,2,3]'. we may do `replace_last(self, ']')'.
changed return type of memory related methods to void.
for a safe malloc to check.
added doc skeleton to it.
added test skeleton to it.