fixup FOLLOW test.
The reason is a left-out of the text book.
During the incremental evaluation of FOLLOW,
1. A := aBb form, if b is epsion or First of b
contains epsion && B is non-terminal, then
union(FOLLOW(B), FOLLOW(A)).
2. A := aB, if B is non-terminal, then
again,
union(FOLLOW(B), FOLLOW(A)).
The text bool left out (2).
But, saddly, heap corruption hit at the string,
which is buggy and unclear.