Encounter a BIG problem in FOLLOW and FIRST
First, `utillib_rule_index_symbol_index' converts
every symbol **defined by client** as opposed to
specail symbol, namely epsilon, eof, error, null,
to an index ranging from 0 to max-1 of their size.
But `eof' is also a terminal symbol which has value
0 and conflicts with other terminals.
This causes wrong FOLLOW and TABLE construction since
`eof' not only comes into FOLLOW but also a column in
TABLE.
Thus, we must (a) eliminate the conflict in the FOLLOW,
(b) give `eof' another column in the TABLE.
to address (a), we can use the `flag' field of `utillib_ll1_set'.
just as how we do with `epsilon' so `eof' will not come into
bitset. In fact, this was supposed to happen.

to address (b) we can let `utillib_rule_index_symbol_index' return
a range [1, max] when passed in a terminal symbol. That is,
preserving the value of `eof' as zero in index.
While this is natural in the parser's point of view (
it needs not treat `eof' specailly but just looks into the table)
 and consistent
with all the treatment of `eof', i.e., the size of terminals counts
`eof', the value of client-defined terminals start from one to preserve
zero, it need extra judgement in the `utillib_rule_index_symbol_index'
and extra documentation.

Another issue is the specail rule `rule_null'.
Maybe we can require `utillib_rule_index' to add the needed
A := epsilon rule instead of using a static place holder, but
that breaks the observation that most code in `utillib_ll1_builder'
does RO operation on `utillib_rule_index'.
But not until `utillib_ll1_builder_build_table' does we know the
needed extra rules, thus the modification is quite limited and
easy to track.

And the benefit is everything can be treated more uniformly and
the parser will be more happy.