Fix another bug caused by treating seq as symbol.
Due to a misunderstanding to the FIRST/FOLLOW algorithm,
previous version of code failed to compute FOLLOW correctly.
To be precise, it calculated the FOLLOW only regarding the
RHS[size] and its predecessor.
However, any adjacent pair of symbols appearing in RHS should
be considered.
The previous commits fix that in one place (inc-eval of FOLLOW)
while this commit fixes the partial eval one.
Worse still, the [here, here+1] loop over the RHS was mistaken
as a nested loop which caused strange FIRST/FOLLOW conflicts.
Luckily, by carefully examination of the code, it was fixed.
One more change. the previous array of function pointers to deal
with rules semantic acttion was isolated because jumping into
a possibly out-of-range array of function pointers is too dangerous.
The current implementaion is one single rule handler where the client
can take different action based on the rule passed in.