Pass cock_rabbit.pas
But notice the program is ended by '.'.
and adding this dot breaks all the things...

Fix bugs in grammar.
The original one is not ll1.
Just a rule of thumb here:
if we are defining a (possibly empty)
list of things, say, a list of idens.
list_of_indens := inden { , inden } | eps
this is a correct EBNF.
now rewrite it to ll1 BNF:
list_of_indens := inden list_of_indens_aux | eps
list_of_indens_aux := , inden list_of_indens_aux | eps
That is it.