Fix up grammar
Now composite stmt is required at the end of
subprogram. since stmt shared the same semicolon
as proc_decl, we can only use begin-and pair to
start the stmt of a subprogram.
this also allows multiple proc_decl in sequence.
but the minimal .pas file now become `begin end.'.
and every stmt in a proc_decl must be wrapped in
begin end;
Notes that proc_decl is separated by semicolon.