Rewrite the hashmap
Also rewrite the slist.
remove all the useless iterator
we are in C and never need such a thing.
The slist used to impl kinda caching of nodes
which is removed as well. we want the struct
to be small and operation to be fast.
good allocator can manage small object quite well
so it is beyond our concern.

hashmap now manages the bucket itself rather than
using a vector.
and it provides no default for string as key.
hash operation should be provided by client always.

Hope this rewrite give birth to lr_builder.
but lalr_builder still another piece of tool.
the graph...