Use more descriptive name for all symbols.
put original symbols for ll1_parser into ll1/
since now that ll1 does not work, we does not need
its symbols with name, value and kind which is only
important for generator.
only their name and value are important.
and UTILLIB_ETAB_INIT does it well.
notes that the cling-scanner changes its output from
now on for the name of symbols.