Fix rd_parser into the new framework.
So what is the new framework?
A fine grained skipping schema.
Basic idea: skip as little tokens as possible
and **always** reach to a stable point to begin
reparsing.
Example:
int i, j=1, k, f[10], w[20];
the `j=1' part is illegal, but the following tokens
are all legal. so we try to skip to `,' the nearest
possible one based on grammar. and see that we success,
we continue in the loop to parse the following tokens.
This will left j undefined however, but that is.
IF the skipping hit eof, we do longjmp as before