Fix up bug in previous lookup-insert model
Previous model is, for each little piece of XX_defs
do a lookup but not insert.
When we assembled a whole XX_decl, we insert it as a whole.
Thanks to the assertion in insert_const, we discover that
example like `const int i=1, i=0;' break us since the first
`i' was not inserted so 2 `i's ended up inserted twice.

The new schema is, in the little piece of XX_defs, we **reserve**
a place in the symbol_table for this name. Notes this reserve
is not like vector_reserve.
And when the whole appears, we do a update to these NULL entry.
In this way, we will detect redefinitions right in time.

Notes for the new schema to work, we need hashmap to verify
existence based on key rather than on value which maybe NULL.
so utillib.hashmap.exist_key was added.
and for cling.symbol_table , exist_name was added similarly.
and update was added in parallel with insert.