Write all the ast_check code
Mainly about expression.
varieties are:
call, subscript, arithmetic, boolean, assign,
factor.
Idea is to evalute the type of the application
of an op on its operands.
Notes this is an ternary algebra since undef is
also a valid result.
so the evaluation rules is designed to be very
tolerant: to avoid returning undef as much as possible