add ast_pretty to the cling.ast family
ast family consists of a group of files
each of which defines a group of functions
that operate on ast and have similar signature.
for example:
ast_check check ast for rd_parser using its symbol_table,
context and input to produce cling_error for it. So it can
simply check its return value against CL_UNDEF to determine
save or discard of this branch of ast.
ast_pretty is pretty like json_tostring but it turn the ast back
into source code, which is useful to print fragment of code to errmsg.
ast_build provide ast creator and setter to wrap those json operation
for different parsers.