Write the local common subexpr elimination (lcse)
Primarily track these kinds of expr: binary, load_lvalue, load_rvalue,
store. Untracked are index, call, ldimme. For ldstr, we always give it
t0. These tracked exprs are represented as lcse_ir.

the idea is to make a hashmap on the lcse_ir with rule to compare different
subclass of expr defined, so that given a lcse_ir we can tell whether
it is in the hashmap or not.
Then we set up other different mapping to help translate ast_ir to
lcse_ir. These are names, values and variables. Names maps a name to
a variable that holds its address (loaded with is_rvalue=false).
Variables maps the temp from ast_ir to the counterpart in lcse_ir, which
is a many-to-one relation. Values maps the variables keeping address of
names to yet another variable that holds its values. This mapping can
change as new value are "store"d.

This algorithm can eliminate common subexprs together with rebundant
loading of value or address and storing of value as it can not only
keep track of which variables are assigned to which temp, but also
updates the value held by each address variables.

The rule for subclasses are listed:

binary only opcode and operands are concerned.

load_lvalue name and scope are concerned. the same name at the same scope
means the same address.

load_rvalue is more complex, as loading the same value twice
is rebundant. Given the name, first we lookup its address from
Names. Then we lookup the value at this address from Values. Finally
we check if Operations gives a value recent enough, if so, we map that
recent value to the temp of load. If not so, we make it a new operation
putting it into the Operations.

store is similar to load_rvalue, as storing the same value to the same
address is rebundant. We first lookup the variable that holds the
address of store from Variables. Then we lookup the value to store from
Variables. These 2 operands is key to check rebundance. However, if we
are storing new thing to an address, Values must be updated to reflect
this. That's why values are not stored together with address in Names:
We need to lookup them using address as keys, not names.
If we forget to update Values, following loads maybe mistakenly omitted.

When the mapping above are maintaining well, there is exception to this
tracking system, which are:

index Although reducing index locating code is wanted, the current
ast_ir couples index locating and value loading in one index ir making
it hard to reuse.
call has arguments pushed before it is issued which harden the tracking
of its rebundance (how to collect those pushes into one lcse_ir?)
ldstr always use t0.
ldimme is never reused since constant can be folded and then propergated
by other algorithms. Treating a single constant as a common expr is
trivial but adds complexity to the system. Not worth.

About reuse of temps
Obviously reusing temps is reusing registers which is good. the temp
whose livetime finishes should be recycled. BUT it is a global
optimization!