New mips schema coined
LRU is abandoned, things should always be easy and code should be
short.
Two level mapping borrowed from lcse is adopted. Names, temps, and regs.
Names holds all the names of local var to temps. One name can be mapped
to lots of temps, which are called aliases of it. A map is signaled by
a load temp name(global=false, rvalue=true) mapping name to temp.
these temps that have a name behind them can have their kind as MIPS_ARG
or MIPS_SAVED and can be put into the saved_regs.
Also the Names map sets up name to their offset.
the Temps map maps from temp to register and maybe their copy on the
stack. All the temps that are not mapped to MIPS_SAVED or MIPS_ARG
will have a copy on the stack.
And the temps that is currently MIPS_TEMP when call happens will be
saved on the stack