Abort the type-check of case
and read_integer has an unsolvable problem:
after shiftaway is called, the new uint will have
been read in so it is imposs to put the nega sign
in front of the uint.
This barrar is due to the limitation of the utillib.scanner
design (lookahead and shiftaway)
one poss workaround is manually modify the internal buffer of
utillib.scanner but it is not elegant
the rd_parser is too deteminant that loses lots of flexibility.
In the program function a SM should have been used but not, which
cause a problem to arise.
const int i=1;
void main() {} will fail.
fix first_ret_function to first_function making it detect main fix it
but too often, it is inelegant.